--- tapShareFrontend/src/components/FireButtonWrap.jsx
@@ -1,164 +1,0 @@
-const FieButtonWrap = () => {
-    return (
-        <svg
-            xmlns="http://www.w3.org/2000/svg"
-            xmlnsBx="https://boxy-svg.com"
-            width="307"
-            height="273"
-            viewBox="42.304 1.633 722 373.528"
-            className="animate-flameMove"
-        >
-            <g bxOrigin="1.154 1.586">
-                <path
-                    fill="#DF513D"
-                    fillRule="evenodd"
-                    d="M217.934 126.101c-1.167-3.763-2.061-7.788-5.236-11.302.108 2.457-.002 4.26-.827 5.933-.684 1.387-.368 3.43-2.745 3.684-2.311.248-3.482-.874-4.668-2.691-3.922-6.005-2.688-12.452-1.678-18.786.745-4.666 2.17-9.221 3.387-14.22-9.078 5.882-13.839 18.679-11.527 29.102 2.305 10.385 6.331 19.888 12.472 28.634 7.29 10.382 7.329 20.787.019 30.697 2.168.269 3.337-.783 4.553-1.723 8.892-6.871 10.305-16.748 10.146-26.877-.118-7.601-1.635-15.158-3.896-22.451zM537.457 199.138c-3.573 3.704-3.719 8.707-4.095 13.078-.443 5.159 2.751 9.729 6.305 13.933 1.678-4.575 1.526-8.778-.152-13.235-1.634-4.335-2.73-8.928-2.058-13.776z"
-                    className="flame one"
-                    clipRule="evenodd"
-                    transform="matrix(.33457 0 .00701 .6479 37.763 11.016)"
-                ></path>
-                <path
-                    fill="#DF513D"
-                    fillRule="evenodd"
-                    d="M790.553 136.011c-1.086-.688-1.059.386-1.111.802-.26 2.063-1.121 4.191.15 6.185 2.043 3.204 3.762 6.5 3.252 11.266 3.506-3.165 4.613-6.646 4.301-10.125-.346-3.828-3.465-6.15-6.592-8.128z"
-                    className="flame two"
-                    clipRule="evenodd"
-                    transform="matrix(.33457 0 .00701 .6479 37.763 11.016)"
-                ></path>
-                <path
-                    fill="#DF513D"
-                    fillRule="evenodd"
-                    d="M939.061 13.063c-2.963-.039-4.814 2.08-4.898 5.601-.365 3.134 2.238 3.978 4.217 4.556 2.504.733 5.953-2.514 5.951-5.005-.001-2.702-2.47-5.114-5.27-5.152z"
-                    className="flame one"
-                    clipRule="evenodd"
-                    transform="matrix(.33457 0 .00701 .6479 37.763 11.016)"
-                ></path>
-                <path
-                    fill="#DF513D"
-                    fillRule="evenodd"
-                    d="M553.012 173.176c-5.986 4.961-6.033 6.817-1.004 11.31 3.383-3.366-.086-7.088 1.004-11.31z"
-                    className="flame"
-                    clipRule="evenodd"
-                    transform="matrix(.33457 0 .00701 .6479 37.763 11.016)"
-                ></path>
-            </g>
-            <path
-                fill="#DF513D"
-                fillRule="evenodd"
-                d="M301.425 459.787c28.333 1.917 57.825-174.448 57.825-174.448 16.786-66.962 33.38-156.46 33.38-156.46-8.147-2.564-14.977-32.723-16.221-45.742-.368-3.835-.083-7.99.413-11.955 1.187-9.477 3.48-18.044 7.76-24.9 1.954-3.131 4.182-5.538 7.532-4.647-4.645-4.46-9.185-5.501-13.976-3.036-4.272 2.2-7.754 6.39-10.93 11.433-3.367 5.347-6.234 11.197-7.988 18.409-.816 3.348-1.498 6.804-1.643 10.318-.19 4.651-.2 9.38.411 14.03.603 4.573 1.518 9.013 2.928 13.032 1.737 4.96 3.551 9.79 4.74 15.221 1.212 5.524 1.777 11.147 1.358 16.884-.5 6.799-2.227 12.75-5.269 17.634-2.505 4.028-5.336 3.804-7.428-.82-.93-2.055-1.584-4.225-1.995-6.814-.876-5.515-.453-10.462 1.616-14.884 2.437-5.206 3.682-10.762 2.744-17.377-.906-6.385-4.913-12.18-8.263-12.319 1.411 1.798 3.447 5.734 4.341 8.404 1.474 4.397 1.293 9.518-1.085 13.068-1.773 2.648-3.34 5.626-4.657 8.973-.993 2.535-1.918 5.086-2.237 8.058-.162 1.516-.036 3.115-.07 4.674-.012.523.119 1.388-.572.914-1.792-1.218-3.226-3.12-3.33-7.326-.057-2.221.67-4.164.693-6.333.056-5.422-2.728-10.85-6.278-11.736-3.262-.816-5.823 1.565-7.535 7.542.785-.438 1.5-.844 2.22-1.24.776-.43 1.584-.164 2.371-.179 2.409-.044 4.782 4.138 4.353 8.277-.366 3.508-1.426 6.497-2.15 9.714-1.294 5.755-3.25 11.068-4.461 16.903-1.271 6.12-2.133 12.288-1.296 18.833.144 1.122.707 2.113.48 3.667-2.522-4.32-5.219-7.916-8.302-10.702-2.716-2.446-5.523-4.51-8.577-5.49-.695-.222-.993-.811-1.136-1.768-.333-2.214-1.154-3.646-2.549-3.904-1.505-.276-2.304 1.431-2.857 3.315-.326 1.121-.65 1.185-1.327 1.036-4.985-1.112-8.695-5.547-11.62-11.917-3.816-8.308-5.444-17.645-4.883-28.015.47-8.71 2.63-16.15 6.073-22.68 3.502-6.646 7.89-11.352 12.74-15.113-.114-.09-.174-.166-.239-.185a.475.475 0 00-.25.008c-6.294 1.911-14.423 7.898-18.81 16.796-.05.106-.113.2-.171.3-1.312 2.224-2.362 4.76-3.227 7.492-1.746 5.54-2.636 11.467-3.133 17.642-.363 4.512-.447 9.044.022 13.43.61 5.693 1.813 11.145 3.678 16.21 1.751 4.752 3.899 9.037 5.725 13.688.708 1.797.65 3.158-.358 4.345a10.042 10.042 0 01-1.975 1.785c-2.212 1.506-4.514 1.627-6.858.972-4.834-1.356-8.493-5.227-10.367-12.907-1.626-6.666-1.272-13.453-.34-20.181 1.093-7.883 2.962-15.413 4.085-23.292 1.014-7.105 1.831-14.27 2.055-21.571.17-5.507.09-11.023-.473-16.494-.363-2.148-.292-4.518-1.269-6.344l.002.001c-.05-2.247-.784-4.089-1.38-6.014-1.113-4.214-2.865-7.529-5.228-9.99.98 2.243 2.011 4.405 2.882 6.762.12.33.295 1.109.762.7-.021 2.198 1.032 3.798 1.246 5.864-.06.246-.071.452.145.48l-.012.045c.437 3.812 1.255 7.505 1.363 11.423.131 3.677-.01 7.32-.275 10.986-.348 4.78-.857 9.479-1.74 14.06-.794 4.134-1.76 8.167-2.678 12.227-1.051 4.653-2.221 9.244-3.148 13.983-1.496 7.635-3.269 15.104-4.224 23.032-1.122 9.301-.573 18.02 2.524 26.188.293.775.44 1.56.479 2.483.267 6.337-1.976 11.29-6.358 11.838-5.284.66-9.458-3.444-10.898-11.262-.738-4.018-.637-8.097-.14-12.158.961-7.902 3.218-14.842 6.01-21.353 1.554-3.623 2.606-7.598 2.74-11.972.28-9.099-1.88-16.558-6.38-22.017-1.67-2.029-3.582-3.335-5.82-2.68-3.788 1.109-4.95 7.755-5.116 13.722.612-1.42 1.177-2.826 1.8-4.161 1.121-2.41 2.745-2.868 4.386-2.264 4.234 1.563 6.268 8.831 4.232 14.942-.803 2.414-1.853 4.502-3.182 6.31-1.615 2.2-2.932 4.991-4.571 7.138-1.869 2.45-3.544 5.227-4.82 8.407-1.64 4.1-3.63 7.352-6.294 9.465-2.291 1.817-5.49 2.318-7.271-1.042-1.527-2.884-1.491-6.238-.82-9.562 1.981-9.84 4.35-19.47 5.658-29.664.737-5.727 1.12-11.488 1.002-17.325-.22-11-2.652-20.728-6.715-29.423-1.222-2.62-2.607-5.04-4.05-7.804-.121 1.633.122 2.773.269 3.83.921 6.646 1.346 13.391 1.642 20.152.176 4.073-.042 8.14-.243 12.229-.29 5.831-.938 11.458-2.107 16.954a45.214 45.214 0 01-3.641 10.541c-1.362 2.78-3.118 5.01-5.083 6.478-2.623 1.96-5.705-.91-6.264-6.013-.952-8.714-.329-17.08 2.825-24.576-.31-.211-.548.044-.836.424-2.363 3.123-4.485 6.52-6.137 10.796-2.874 7.449-4.167 15.69-4.807 24.237-.49 6.537-1.029 13.013-2.234 19.318-1.422 7.447-6.601 12.796-11.356 11.631-3.015-.739-5.105-3.749-6.698-7.66-4.681-11.501-4.205-23.369-.804-35.334 1.286-4.527 3.099-8.514 5.066-12.44-4.03-.8-7.228 1.595-9.881 5.98-3.196 5.288-4.551 12.202-5.827 19.084-.296 1.598-.55 1.168-1.04.55-4.386-5.524-7.216-12.666-8.392-21.508-.178-1.61-.355-3.221-.535-4.831-.038-6.11.38-12.108 1.596-17.96-.311 1.006-.912 1.47-1.264 2.352-4.242 10.678-4.934 22.537-4.261 34.808.149 2.755-.119 5.305-.643 7.795-.887 4.23-2.55 5.05-4.676 2.246-1.469-1.936-2.837-4.103-4.135-6.345-3.858-6.66-5.969-14.556-6.286-23.734-.36-10.396 1.172-20.197 3.473-29.7 2.633-10.871 3.384-21.898 1.61-33.339a62.156 62.156 0 00-5.286-17.17c-1.261-2.659-2.909-5.024-4.748-6.886-5.266-5.318-10.615-6.527-16.258-1.41.477.572.905.825 1.399 1.029 3.863 1.602 7.375 4.483 10.171 9.104 3.385 5.592 4.268 12.61 3.41 20.169-.805 7.087-2.228 13.914-3.884 20.555-2.354 9.447-5.19 18.567-6.83 28.483-.747 4.503-1.224 9.036-1.316 13.693-.008.385-.017.942-.161 1.108-.323.374-.535-.21-.704-.541a25.375 25.375 0 01-2.067-5.696c-1.758-7.493-1.34-15.01-.067-22.466.679-3.964 1.616-7.807 2.553-12.265-1.553 1.755-2.268 3.94-3.025 5.982-3.424 9.259-4.022 19.439-3.16 29.888.605 7.327 7.085 27.08 1.977 21.958-8.657-8.683-31.78-44.456-18.52 68.212 0 0 36.017 255.86 123.899 249.197 4.165-.315 97.971-22.025 97.971-22.025.672-.016-80.547-196.905-79.878-196.937l8.339-.203 1.431-.041c1.054-1.017 1.873-2.248 2.507-3.65-.633 1.402-1.453 2.633-2.507 3.65l5.826-.13c.461.008.921.037 1.383.027 23.928-.553 47.856-1.107 71.785-1.67 8.73-.206-36.114 131.394-27.387 131.174 5.353-.127 11.246 29.146 11.708 29.136m47.53-286.177l.001.001-.015.303.014-.304zm-80.575-45.249c.326.43.493 1.169.817 1.599-.323-.428-.491-1.167-.817-1.599zM106.436 197.82l.551-3.547c-.015-.025-.551 3.546-.551 3.547z"
-                className="flame-main one"
-                clipRule="evenodd"
-                bxOrigin="1.038 0.471"
-            ></path>
-            <path
-                fill="#F26C52"
-                d="M334.548 222.403c.295 2.707.497 5.614.596 8.759 0 0-.572-56.114-16.303-89.894-2.785-7.566-4.981-16.702-6.644-27.204-1.187-7.504-1.958-15.451-2.058-23.814-.146-12.33.987-23.36 3.079-33.236.964-4.548 2.026-8.816 3.369-12.225.171-.436.3-1.052.565-2.006-1.393.694-2.481 2.505-3.546 4.298-2.577 4.339-4.596 10.772-6.074 18.944-1.355 7.489-1.976 15.828-1.999 24.645-.044 15.678 1.56 29.969 3.925 43.556 1.206 6.923 2.538 13.643 3.017 21.622.905 15.039.592 29.256-1.988 42.111-1.789 8.907-5.542 14.044-8.625 11.559-2.061-1.663-3.178-6.691-3.896-12.724-1.357-11.409-1.163-22.31.865-32.404.645-3.209 1.268-6.467 1.94-9.911-.981.834-3.384 10.512-4.227 16.706-.963 7.068-.195 14.669-.389 21.973-2.304-2.338-5.188-10.327-4.972-19.487.083-3.52-.047-7.067.199-10.604.3-4.294-.337-8.278-1.527-10.501-.768-1.434-1.584-2.428-2.479-2.245.661 1.142 1.391 1.68 1.948 3.539.414 1.378.658 2.887.696 4.739.197 9.907-.144 19.262-1.587 28.147-1.8 11.081-2.652 23.218-2.321 36.146.125 4.896.654 9.367 1.719 13.062 1.873 6.495 2.321 14.628 1.384 22.49-.15 1.259-.277 2.458-.844 1.128-2.59-6.092-5.15-12.314-7.378-19.869-.986-3.341-1.829-7.139-2.813-10.49-1.801-6.14-4.079-8.447-6.55-9.601-3.669-1.713-7.336-3.627-10.443-11.701-2.021-5.249-3.38-12.005-4.431-19.5-1.318-9.408-2.138-19.162-2.123-29.529.009-6.557.359-12.802.999-18.884.043-.409.083-.826.165-1.653-.444.532-.493 1.595-.59 2.418-.922 7.888-1.64 16.012-1.951 24.464-.198 5.298.049 10.62.564 15.87.932 9.522 2.489 17.947 4.194 26.159.799 3.852 1.603 7.698 2.368 11.624 1.22 6.258.597 13.767-1.334 17.284-1.825 3.324-3.783 3.818-5.831 3.003-.136-.055-.276.058-.415.093l-.367-.022c-1-1.231-2.011-2.386-3.002-3.706-3.973-5.292-6.789-17.113-6.975-31.846-.117-9.211.498-17.802 1.201-26.41.733-8.976 1.499-17.919 2.232-26.895.136-1.676.183-3.446.082-5.382-.376 3.452-.769 6.886-1.126 10.365-1.003 9.754-2.053 19.46-2.968 29.317-.823 8.858-1.642 17.821-1.503 27.348.137 9.257 1.119 17.614 2.865 25.156.751 3.245 1.774 5.843 2.261 9.716.915 7.27.42 13.857-.79 19.94-2.056 10.341-5.108 12.786-8.718 11.304-1.605-.659-3.14-2.016-4.641-3.882-2.973-3.697-4.498-11.712-5.39-21.085-1.382-14.539-.658-27.979 1.72-40.466 1.506-7.911 3.313-15.04 5.008-22.467 2.084-9.138 1.738-20.475-.823-28.818-1.142-3.718-2.247-7.444-3.939-8.226-.695-.32-1.314-.238-1.904 1.128 1.492-.108 2.75 1.824 3.817 5.046 2.684 8.107 3.067 17.843 1.248 28.234-.987 5.642-2.187 10.755-3.526 15.392-1.97 6.827-3.682 14.342-5.547 21.437-1.462 5.556-3.215 10.117-5.528 11.352-2.932 1.565-6.42-3.186-6.819-15.685-.292-9.143.323-17.615 1.289-25.825.794-6.744 1.765-13.236 2.598-19.928 1.15-9.254 1.203-18.91.663-28.874-.39-7.175-1.039-14.108-1.651-21.091-.058-.651-.126-1.26-.314-1.647-.276.278-.175 1.056-.175 1.65-.014 11.724-.411 23.276-1.07 34.707-.507 8.807-1.163 17.439-2.372 25.47-1.265 8.406-3.161 14.045-6.026 16.085-1.979 1.409-3.313-1.799-4.08-6.921-1.223-8.189-1.487-16.998-1.243-25.902.048-1.749.09-3.501.154-6.051-.778 5.352-1.244 10.264-1.663 15.23-.833 9.881-.982 20.153-1.286 30.352-.256 8.564-.581 17.13-1.564 25.089l-3.274 14.72c-1.558 3.76-3.479 5.524-5.756 5.222-1.871-.247-3.443-3.472-4.672-8.224-2.557-9.891-3.878-21.535-4.502-34.024-.541-10.808.225-20.978 1.253-30.955.349-3.384.925-6.385 1.444-9.527-1.301 3.218-2.412 6.982-2.971 12.125-.626 5.753-1.13 11.662-1.752 17.419-.319 2.939-.644 6.408-1.756 6.789-1.042.359-1.674-2.515-2.294-4.929-1.064-4.156-2.122-8.354-2.956-13.044-1.188-6.669-1.752-13.949-2.063-21.496-.114-2.794-.203-5.586-.309-8.386-.351 10.637-.467 21.265.192 31.973.26 4.235.446 8.504.261 12.863-.447 10.549-2.92 14.109-5.802 11.249-3.798-3.773-6.254-12.475-7.891-24.067-1.858-13.156-2.458-27.024-2.31-41.114.109-10.389.878-20.36 1.8-30.248.832-8.944 1.925-17.542 2.725-26.514.974-10.943.331-21.399-1.025-31.793-.663-5.087-1.51-9.627-2.818-13.2-.293-.8-.569-1.586-1.076-1.686.428 3.604.896 6.941 1.224 10.407 1.293 13.664.822 26.785-.862 39.383-.847 6.342-1.846 12.448-2.7 18.785-1.913 14.228-2.951 29.314-3.526 44.87-.251 6.83-.382 13.668-.26 21.437-1.999-7.434-3.749-14.399-5.329-21.939.291 4.354.629 8.686.861 13.067.56 10.564 1.213 21.121.617 31.825 0 0-3.745-8.631-6.923 26.485a505.963 505.963 0 00-.774 9.608c1.636-3.968 2.298-9.278 2.296-15.843 0-.446.021-.891.047-1.93.398.965.716 2.01.965 3.134-29.172 63.403 42.076 255.905 64.536 244.371 34.056 10.138 47.612-16.052 73.753-14.825.382.015-83.41-52.335-83.03-52.319 32.184 1.4 97.097 15.881 115.121 11.769.229-.055-25.04 33.002-24.813 32.948 19.251-4.457 23.725 8.582 23.725 8.582 37.465-9.334 55.954-204.3 32.242-233.014z"
-                className="flame-main two"
-                bxOrigin="1.275 0.459"
-            ></path>
-            <path
-                fill="#F58553"
-                fillRule="evenodd"
-                d="M306.688 451.925c1.057 0 25.925-86.793 26.98-86.852-6.011-12.569-6.443-32.308-6.287-50.629-.799.285-1.099 2.577-1.55 4.105-2.165 7.351-3.017 17.739-4.576 26.498-2.026 11.384-3.748 11.49-5.9-.026-.628-3.349-.998-3.247-1.818-1.438-2.314 5.094-4.376 4.065-6.472-2.832-1.445-4.76-2.418-10.886-3.669-16.183-1.121-4.746-1.468-9.817-1.613-15.673-.182-7.38.322-14.487.189-21.791-1.695 4.243-2.217 11.178-2.971 17.52-.919 7.74-1.13 16.284-1.792 24.303-.53 6.414-1.635 11.107-3.807 10.669-1.786-.363-3.452-5.211-4.04-12.164-.23-2.724-.357-5.549-.542-8.322-.64-9.554-2.748-16.833-5.511-18.756.372 2.433.75 4.593 1.045 6.882 1.1 8.556.172 18.08-2.131 22.156-2.102 3.717-4.564.439-5.839-8.287-1.542-10.568-2.164-21.36-.828-32.684 1.54-13.053 3.292-25.835 3.988-40.08.709-14.5 1.553-28.982.221-43.862-.39-4.362-.789-8.857-1.997-11.639.307 17.177.068 33.717-1.412 49.615-.328 3.531-.796 7.622-2.026 7.431-1.195-.183-1.722-4.298-2.114-7.852-.713-6.478-1.051-13.289-1.361-20.226-.996 6.802-1.459 13.881-1.518 21.543-.16 20.399 1.498 40.485 1.437 60.806-.039 13.104-.64 25.736-3.31 35.61-1.238 4.575-2.741 6.26-4.455 5.349-1.492-.793-1.186-6.295-1.661-9.622-1.154-8.101-2.661-14.385-5.169-16.806.082 1.931.312 3.335.548 4.723 1.679 9.848.978 20.396-1.79 26.926-1.722 4.066-3.681 1.585-5.717-7.624-1.502-6.786-2.76-14.341-4.406-20.606-2.204-8.383-2.557-17.667-2.407-28.236.116-8.16 12.441-13.932 12.57-22.492-.671.978-33.429-18.351-33.606-16.948-1.131 8.959 19.32 37.682 18.486 46.957-.474 5.29-.836 10.845-2.27 14.186-1.275 2.974-2.494-.104-4.008-1.423.497 4.023.52 7.269.557 10.525.066 5.849-.673 9.995-2.078 12.359-1.154 1.944-2.452 2.592-3.563-.584-1.279-3.643-1.258-8.849-.927-13.771.513-7.624.595-15.367.338-23.156-.304-9.274-1.174-17.921-2.869-27.118-.223 20.173-2.572 35.539-4.867 50.865-1.649 11.016-4.389 17.377-7.815 20.098-5.033 3.987-10.322-6.061-12.293-23.263-.461-4.02-.798-8.436-2.412-11.826.815 5.075.713 9.115.737 13.079.057 9.508-2.341 15.153-4.84 11.464-2.157-3.185-2.546-10.061-2.722-17.001-.419-16.585.197-32.922.473-49.35.236-13.962.067-27.91-1.633-41.151-.6-4.667-1.103-9.796-2.716-13.019.331 5.22.758 9.702.879 14.253.461 17.213-.504 33.733-1.615 50.105-.339 4.99-1.3 7.579-2.851 7.819-4.34.659-7.801-7.609-10.58-17.649-2.855-10.314-4.779-23.699-5.661-38.317-.644-10.682-1.472-21.202-2.632-31.427-1.604-14.139-3.077-21.894-5.553-27.498 3.067 17.531 4.565 35.283 4.338 54.477-.014 1.123.145 2.304.234 3.449 1.688 21.921 4.893 41.329 8.33 60.34.983 5.442 1.564 11.366 1.292 17.575-.609 13.922-2.408 25.499-5.481 34.245-.023.189-.037.413-.073.555-.021.083-.082.033-.125.044l.199-.598c-.034-8.779-.013-17.611-2.053-25.036-.632 8.851-1.625 13.991-4.952 25.121 0 0 13.508 101.808 21.3 111.857 12.697 16.377 51.118 20.52 124.743.308z"
-                className="flame-main three"
-                clipRule="evenodd"
-                opacity="0.8"
-                bxOrigin="1.411 0.281"
-            ></path>
-            <g bxOrigin="1.202 1.37">
-                <path
-                    fill="#F58553"
-                    fillRule="evenodd"
-                    d="M260.138 279.034c.329 2.103.929 3.955 3.466 1.591 1.36-1.269 2.555-2.34 2.946-4.48.611-3.344 1.288-6.88 4.965-9.637-8.724.601-12.534 5.132-11.377 12.526z"
-                    className="flame"
-                    clipRule="evenodd"
-                    transform="matrix(.44015 0 .00757 .69964 10.112 -6.053)"
-                ></path>
-                <path
-                    fill="#F58553"
-                    fillRule="evenodd"
-                    d="M642.133 261.121c-.602 1.805 2.854 4.751 5.137 4.486 2.775-.322 5.049-1.429 4.986-4.831-.051-2.835-2.447-5.298-5.188-5.287-3.64.102-4.129 3.208-4.935 5.632z"
-                    className="flame one"
-                    clipRule="evenodd"
-                    transform="matrix(.44015 0 .00757 .69964 10.112 -6.053)"
-                ></path>
-                <path
-                    fill="#F58553"
-                    fillRule="evenodd"
-                    d="M236.169 192.895c2.469-.638 4.981-.998 4.781-3.98-.117-1.744-.676-3.642-3.098-3.758-2.766-.133-4.256 1.769-4.511 3.915-.178 1.502 1.072 3.33 2.828 3.823z"
-                    className="flame two"
-                    clipRule="evenodd"
-                    transform="matrix(.44015 0 .00757 .69964 10.112 -6.053)"
-                ></path>
-                <path
-                    fill="#F58553"
-                    fillRule="evenodd"
-                    d="M394.363 104.625c2.114.205 3.56-.855 3.625-2.719.057-1.631-1.206-2.715-3.106-2.809-1.935-.095-2.961.578-3.069 2.6-.105 1.918.485 3.084 2.55 2.928z"
-                    className="flame"
-                    clipRule="evenodd"
-                    transform="matrix(.44015 0 .00757 .69964 10.112 -6.053)"
-                ></path>
-                <path
-                    fill="#F58553"
-                    fillRule="evenodd"
-                    d="M257.108 216.734c1.575.05 2.945-.246 2.794-2.009-.133-1.558-1.21-2.582-2.89-2.516-1.492.059-2.595 1.087-2.394 2.435.156 1.042.819 2.58 2.49 2.09z"
-                    className="flame one"
-                    clipRule="evenodd"
-                    transform="matrix(.44015 0 .00757 .69964 10.112 -6.053)"
-                ></path>
-                <path
-                    fill="#F58553"
-                    fillRule="evenodd"
-                    d="M73.648 152.806c1.225.057 1.942-.5 2.374-1.896-.912-.418-.55-1.965-2.227-2.114-1.723-.152-2.062 1.195-2.287 2.05-.389 1.471.828 1.898 2.14 1.96z"
-                    className="flame two"
-                    clipRule="evenodd"
-                    transform="matrix(.44015 0 .00757 .69964 10.112 -6.053)"
-                ></path>
-            </g>
-            <g bxOrigin="0.305 1.624">
-                <path
-                    fill="#DF513D"
-                    fillRule="evenodd"
-                    d="M-858.734 168.791c-1.167-3.763-2.061-7.788-5.236-11.302.108 2.457-.002 4.26-.827 5.933-.684 1.387-.368 3.43-2.745 3.684-2.311.248-3.482-.874-4.668-2.691-3.922-6.005-2.688-12.452-1.678-18.786.745-4.666 2.17-9.221 3.387-14.22-9.078 5.882-13.839 18.679-11.527 29.102 2.305 10.385 6.331 19.888 12.472 28.634 7.29 10.382 7.329 20.787.019 30.697 2.168.269 3.337-.783 4.553-1.723 8.892-6.871 10.305-16.748 10.146-26.877-.118-7.601-1.635-15.158-3.896-22.451zM-409.033 191.2c-3.573 3.704-3.719 8.707-4.095 13.078-.443 5.159 2.751 9.729 6.305 13.933 1.678-4.575 1.526-8.778-.152-13.235-1.634-4.335-2.73-8.928-2.058-13.776z"
-                    className="flame one"
-                    clipRule="evenodd"
-                    transform="matrix(.4087 0 .00763 .70505 481.812 -7.577)"
-                ></path>
-                <path
-                    fill="#DF513D"
-                    fillRule="evenodd"
-                    d="M790.553 136.011c-1.086-.688-1.059.386-1.111.802-.26 2.063-1.121 4.191.15 6.185 2.043 3.204 3.762 6.5 3.252 11.266 3.506-3.165 4.613-6.646 4.301-10.125-.346-3.828-3.465-6.15-6.592-8.128z"
-                    className="flame two"
-                    clipRule="evenodd"
-                    transform="matrix(.4087 0 .00763 .70505 481.812 -7.577)"
-                ></path>
-                <path
-                    fill="#DF513D"
-                    fillRule="evenodd"
-                    d="M939.061 13.063c-2.963-.039-4.814 2.08-4.898 5.601-.365 3.134 2.238 3.978 4.217 4.556 2.504.733 5.953-2.514 5.951-5.005-.001-2.702-2.47-5.114-5.27-5.152z"
-                    className="flame one"
-                    clipRule="evenodd"
-                    transform="matrix(.4087 0 .00763 .70505 481.812 -7.577)"
-                ></path>
-                <path
-                    fill="#DF513D"
-                    fillRule="evenodd"
-                    d="M553.012 173.176c-5.986 4.961-6.033 6.817-1.004 11.31 3.383-3.366-.086-7.088 1.004-11.31z"
-                    className="flame"
-                    clipRule="evenodd"
-                    transform="matrix(.4087 0 .00763 .70505 481.812 -7.577)"
-                ></path>
-            </g>
-            <path
-                fill="#F58553"
-                fillRule="evenodd"
-                d="M327.165 422.934c-15.221-92.081-114.768-82.94-125.803-2.889-.164 1.189-37.126-75.968-38.229-76.033 3.459-6.213 32.281-5.525 33.369-15.765.72 1.004 1.049 2.39 1.51 3.543 7.317 18.288 16.011 15.701 23.569 4.36 2.962-4.444 5.035-10.125 3.877-19.208 3.703 1.273 6.233 4.93 7.777 12.067 1.443 6.66 3.521 7.138 5.737 1.711 1.015-2.49 1.386-5.407 1.481-8.789.12-4.245-1.123-7.511-1.857-11.518 1.666.424 2.562 2.176 3.468 3.893 2.947 5.59 5.026 12.407 6.436 20.208 1.403 7.755 3.812 13.552 7.713 15.079 8.43 3.295 19.337-8.529 20.89-29.659.912-12.417 2.236-24.71 5.105-35.793 1.347-5.204 2.69-10.6 5.891-13.76-.662 3.724-1.209 6.413-1.625 9.196-1.425 9.517-2.184 19.053.23 28.656 2.133 8.492 5.9 9.928 9.174 3.563 3.012-5.862 3.428-12.785 1.165-21.501 4.378 2.494 6.496 7.694 7.74 15.065.91 5.397 2.198 6.351 4.54 3.594 2.323-2.735 3.502-7.389 4.237-12.596a163.252 163.252 0 001.541-26.542c-.11-4.897-.125-9.827-1.818-13.833a531.16 531.16 0 01-.578-2.566c5.081 4.401 6.337 14.451 7.738 24.08.833 5.719 1.193 11.731 1.935 17.516.307 2.388.766 5.338 2.147 5.795 1.399.463 2.2-2.097 2.7-4.267 3.551-15.398 4.708-31.059.226-47.072-2.838-16.026-2.517-31.482 1.327-46.301l-.035-.085c.246 8.454-.883 16.764.078 25.385.806 7.218 2.023 13.805 3.958 19.955 1.577 5.014 2.819 10.402 3.428 16.338l.689 8.188.046 11.257c-.937 7.122-2.02 14.166-2.765 21.387-.574 5.579.161 10.744 2.591 14.1 2.352 3.252-7.418 101.269-9.603 103.241z"
-                className="flame-main three"
-                clipRule="evenodd"
-                opacity="0.8"
-                bxOrigin="1.506 0.143"
-            ></path>
-        </svg>
-    )
-}
-export default FieButtonWrap
