+++ tapShareFrontend/src/components/animated/uploadingAnimation.jsx
@@ -1,4 +1,4 @@
-import LoadingSvg from "./loadingSvg";
+import LoadingSvg from "../svg/loadingSvg";
 
 export default function UploadingAnimation() {
   return (
@@ -9,7 +9,7 @@ export default function UploadingAnimation() {
     >
       <div className="flex justify-center items-center px-4 py-2 select-none">
         <div className="h-full flex items-center relative -right-1">
-          Uploading{" "}
+          Uploading
         </div>
         <div className="flex items-center max-w-fit relative right-6">
           <object>
