+++ tapShareFrontend/src/components/misc/appBar.jsx
