+++ tapShareFrontend/src/components/svg/loadingSvg.jsx
