--- tapShareFrontend/src/components/uploadCodeButton.jsx
@@ -1,5 +1,0 @@
-export default function UploadCodeButton() {
-  return (
-    <div>uploadCodeButton</div>
-  )
-}
